requirements.txt
# Discord 봇 (voice 녹음/sinks 지원 - py-cord 필요)
py-cord[voice]>=2.6.0
PyNaCl>=1.5.0

# 음성 전사 (오픈소스 Whisper)
faster-whisper>=1.0.0

# 설정 파일
pyyaml>=6.0.1
python-dotenv>=1.0.0
